fix: recordslistを切り出した
